moved utlility functions
