+++ docs/reference/util.rst
@@ -1,0 +1,4 @@
+Utilities
+=========
+
+.. automodapi:: roentgen.util
